requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlmodel==0.0.21
pydantic-settings==2.4.0
fastapi==0.111.0
pydantic-settings==2.4.0
sqlmodel==0.0.21
uvicorn[standard]==0.30.1
passlib[bcrypt]
python-jose[cryptography]